% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{add_date_and_format}
\alias{add_date_and_format}
\title{Add date and format to activity counts}
\usage{
add_date_and_format(counts)
}
\arguments{
\item{counts}{a data frame containing activity counts}
}
\value{
a data frame with time stamps added in POSIXct format
}
\description{
This function takes a data frame of activity counts and adds a column of time stamps in POSIXct format.
The time stamps start at "2012-04-07 00:00:30" and increase by 30 seconds for each row of the data frame.
}
\examples{
data <- data.frame(steps = c(100, 200, 300))
data_with_time <- add_date_and_format(data)
head(data_with_time)

}
