% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gps_data_into_gps_epochs.R
\name{assign_epoch_start_time}
\alias{assign_epoch_start_time}
\title{Assign Epoch Start Time}
\usage{
assign_epoch_start_time(gps_data, epoch_length)
}
\arguments{
\item{gps_data}{A data frame with GPS data including a column of timestamps and columns for latitude and longitude}

\item{epoch_length}{The duration of an epoch in seconds}
}
\value{
A data frame of GPS data with an additional column indicating epoch start time
}
\description{
Assign Epoch Start Time
}
\details{
Selects the closest 30 second increment to assign epoch start time and takes the GPS coordinates associated with the latest time if there are multiple GPS data points in a given 30 second increment. This function returns a data frame of GPS data with a column of epoch times.
}
