% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{collate_arguments}
\alias{collate_arguments}
\title{Collate Arguments
This function collates user-provided arguments with pre-defined parameters and constants.}
\usage{
collate_arguments(..., collated_arguments = NULL)
}
\arguments{
\item{...}{named arguments passed by the user}

\item{collated_arguments}{NULL or previously collated arguments}
}
\value{
A list of all arguments, including both pre-defined parameters and constants and any user-provided arguments.
}
\description{
Collate Arguments
This function collates user-provided arguments with pre-defined parameters and constants.
}
