% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{generate_gps_data}
\alias{generate_gps_data}
\title{Generate a dataset with date-time, speed, and latitude and longitude of someone moving through space on a walk in Seattle}
\usage{
generate_gps_data(
  start_lat,
  start_long,
  start_time,
  n_epochs = 100,
  time_interval = 30
)
}
\arguments{
\item{start_lat}{The starting latitude of the walk.}

\item{start_time}{The start time of a series of data}

\item{n_epochs}{The number of epochs in the series}

\item{time_interval}{The time interval between points in seconds.}

\item{start_lon}{The starting longitude of the walk.}
}
\value{
A data frame with four columns: "timestamp", "lat", "lon", and "speed".
}
\description{
Generate a dataset with date-time, speed, and latitude and longitude of someone moving through space on a walk in Seattle
}
