% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/identify_walk_bouts_in_gps_and_accelerometry_data.R
\name{identify_walk_bouts_in_gps_and_accelerometry_data}
\alias{identify_walk_bouts_in_gps_and_accelerometry_data}
\title{Identify walking bouts in GPS and accelerometry data:}
\usage{
identify_walk_bouts_in_gps_and_accelerometry_data(
  gps_data,
  accelerometry_counts,
  ...,
  collated_arguments = NULL
)
}
\arguments{
\item{gps_data}{A data frame containing GPS data}

\item{accelerometry_counts}{A data frame containing accelerometry counts}

\item{...}{Additional arguments to be passed to the function}

\item{collated_arguments}{A list of collated arguments}
}
\value{
A data frame containing identified walk bouts
}
\description{
This function identifies walking bouts in GPS and accelerometry data.
It processes the GPS data and accelerometry counts to create walk bouts.
}
