% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{make_active_period}
\alias{make_active_period}
\title{Generate accelerometry datasets}
\usage{
make_active_period(
  length = 1,
  is_bout = TRUE,
  non_wearing = FALSE,
  complete_day = FALSE
)
}
\arguments{
\item{length}{Length of the active period}

\item{is_bout}{Logical indicating if the active period is a bout}

\item{non_wearing}{Logical indicating if the active period is a non-wearing period}

\item{complete_day}{Logical indicating if the active period is a complete day}
}
\value{
A list of activity epochs
}
\description{
This function generates a list of activity epochs with specified minimum active counts per epoch, minimum bout length,
maximum number of consecutive inactive epochs in a bout, minimum non-wearing length, and minimum complete day length.
}
