% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{make_full_day_bout}
\alias{make_full_day_bout}
\title{Create activity counts for a full day bout}
\usage{
make_full_day_bout()
}
\value{
A data frame with activity counts for a full day bout
}
\description{
This function creates a data frame with activity counts for a full day bout. A full day bout is defined as an uninterrupted period of activity with a length of at least \code{min_complete_day}. The function calls the \code{make_non_bout_window()}, \code{make_smallest_bout_window()}, and \code{make_smallest_complete_day_activity()} functions to generate the activity counts for the non-bout window, smallest bout window, and smallest complete day activity, respectively.
}
