% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{make_full_walk_bout_df}
\alias{make_full_walk_bout_df}
\title{Create a data frame of walking bouts with GPS data}
\usage{
make_full_walk_bout_df()
}
\value{
A data frame of walking bouts with GPS data
}
\description{
This function combines accelerometer and GPS data to create a data frame of walking bouts.
It generates a full day of activity with bouts of minimum and non-bout periods, and GPS data for walking in Seattle.
The accelerometer data is processed into bouts using the \code{\link{process_accelerometry_counts_into_bouts}} function.
The GPS data is processed into epochs using the \code{\link{process_gps_data_into_gps_epochs}} function.
}
\examples{
make_full_walk_bout_df()
}
