% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{make_inactive_period}
\alias{make_inactive_period}
\title{Create an inactive period}
\usage{
make_inactive_period(
  length = 1,
  is_bout = FALSE,
  non_wearing = FALSE,
  complete_day = FALSE
)
}
\arguments{
\item{length}{The length of the inactive period.}

\item{is_bout}{Logical value indicating whether this period is part of a bout of inactivity.}

\item{non_wearing}{Logical value indicating whether this period is due to non-wearing of the accelerometer.}

\item{complete_day}{Logical value indicating whether this period occurs during a complete day of wearing the accelerometer.}
}
\value{
A data frame with columns activity_counts, bout, non_wearing, and complete_day, where activity_counts is set to 0 for the entire length, and bout, non_wearing, and complete_day are set according to the input values.
}
\description{
This function creates an inactive period with a given length.
}
\examples{
make_inactive_period(length = 30, is_bout = TRUE, non_wearing = FALSE, complete_day = TRUE)

}
