% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{make_non_bout_window}
\alias{make_non_bout_window}
\title{Create a non-bout window}
\usage{
make_non_bout_window(maximum_number_consec_inactive_epochs_in_bout = 3)
}
\value{
a data frame with columns "activity_counts", "bout", "non_wearing", "complete_day"
}
\description{
This function creates a non-bout window, which is a period of inactivity that is not long enough to be considered as an inactive bout.
}
\examples{
make_non_bout_window()

}
