% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{make_smallest_bout}
\alias{make_smallest_bout}
\title{Make the smallest bout dataset}
\usage{
make_smallest_bout()
}
\value{
A data frame containing the activity counts and bout information for the smallest bout.
}
\description{
Generates a dataset representing the smallest bout, consisting of a sequence of inactive periods followed by the smallest active period.
}
\examples{
make_smallest_bout()
}
