% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{make_smallest_bout_with_smallest_non_wearing_period}
\alias{make_smallest_bout_with_smallest_non_wearing_period}
\title{Generate the smallest bout with the smallest non-wearing period dataset}
\usage{
make_smallest_bout_with_smallest_non_wearing_period()
}
\value{
A data frame with columns for activity counts and date-time stamps.
}
\description{
This function creates a dataset consisting of the smallest bout and the smallest non-wearing period. The bout length, non-wearing period length, and epoch length are defined in the global variables: minimum_bout_length, maximum_number_consec_inactive_epochs_in_bout, and min_non_wearing_length, respectively.
}
\examples{
make_smallest_bout_with_smallest_non_wearing_period()
}
