% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{make_smallest_bout_without_metadata}
\alias{make_smallest_bout_without_metadata}
\title{Create the smallest bout window without metadata}
\usage{
make_smallest_bout_without_metadata()
}
\value{
A data frame containing the smallest bout window without metadata.
}
\description{
This function creates the smallest bout window without the metadata columns. It calls the \code{\link{make_smallest_bout}} function and then removes the columns "non_wearing", "complete_day", and "bout" using \code{dplyr::select}.
}
\examples{
make_smallest_bout_without_metadata()
}
