% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{make_smallest_complete_day_activity}
\alias{make_smallest_complete_day_activity}
\title{Generate an activity sequence for a complete day with minimal activity}
\usage{
make_smallest_complete_day_activity(min_complete_day = 8602)
}
\value{
An activity sequence data frame with minimum activity counts for a complete day.
}
\description{
This function generates an activity sequence for a complete day with a minimal activity count.
}
\examples{
make_smallest_complete_day_activity()

}
