% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{make_smallest_nonwearing_window}
\alias{make_smallest_nonwearing_window}
\title{Create smallest non-wearing window}
\usage{
make_smallest_nonwearing_window(min_non_wearing_length = 20 * 2)
}
\value{
An inactive period data frame that represents the smallest non-wearing window.
}
\description{
Create an inactive period that represents the smallest non-wearing window.
This function uses the \code{make_inactive_period()} function to create the non-wearing window.
}
\examples{
make_smallest_nonwearing_window()
}
