% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{next_lat_long}
\alias{next_lat_long}
\title{Gets the next lat/long coordinate}
\usage{
next_lat_long(latitude, longitude, speed, direction, dt)
}
\arguments{
\item{speed}{The speed of the person in meters per second.}

\item{direction}{The initial direction of the person's movement in degrees.}
}
\description{
Gets the next lat/long coordinate
}
