% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bouts_and_gps_epochs_into_walkbouts.R
\name{outlier_gps_points}
\alias{outlier_gps_points}
\title{Outlier GPS data points
This function identifies outlier GPS points for the bout radius calculation from a given set of latitude and longitude coordinates.}
\usage{
outlier_gps_points(lat_long, dwellbout_radii_quantile)
}
\arguments{
\item{lat_long}{A data frame containing the latitude and longitude coordinates for the GPS points.}

\item{dwellbout_radii_quantile}{The threshold for outliering GPS data points - any data points above the specified percentile are outliered.}
}
\value{
A data frame containing the latitude and longitude coordinates for the non-outlier GPS points.
}
\description{
Outlier GPS data points
This function identifies outlier GPS points for the bout radius calculation from a given set of latitude and longitude coordinates.
}
