% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{Global parameters and constants}
\format{
An object of class \code{list} of length 5.
}
\usage{
parameters
}
\description{
Global parameters and constants
}
\keyword{datasets}
