% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accelerometry_counts_into_bouts.R
\name{process_accelerometry_counts_into_bouts}
\alias{process_accelerometry_counts_into_bouts}
\title{Convert accelerometry counts into bouts of physical activity}
\usage{
process_accelerometry_counts_into_bouts(accelerometry_counts)
}
\arguments{
\item{accelerometry_counts}{Accelerometry data}
}
\value{
A data frame of bouts.
}
\description{
Convert accelerometry counts into bouts of physical activity
}
\examples{
x <- process_accelerometry_counts_into_bouts(acc_data)
}
