% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/process_bouts_and_travel_instances_into_walkbouts.R
\name{process_bouts_and_travel_instances_into_walkbouts}
\alias{process_bouts_and_travel_instances_into_walkbouts}
\title{Convert bouts and travel instances into walk bouts}
\usage{
process_bouts_and_travel_instances_into_walkbouts(bouts, travel_instances)
}
\arguments{
\item{bouts}{Accelerometry bouts}

\item{travel_instances}{GPS travel instances}
}
\value{
A data frame of walk bouts and summary statistics.
}
\description{
Convert bouts and travel instances into walk bouts
}
\examples{
x <- process_bouts_and_travel_instances_into_walkbouts(bouts, travel_instances)
}
