% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/process_gps_and_accelerometry_data_into_walkbouts.R
\name{process_gps_and_accelerometry_data_into_walkbouts}
\alias{process_gps_and_accelerometry_data_into_walkbouts}
\title{Convert accelerometry and GPS data into walk bouts}
\usage{
process_gps_and_accelerometry_data_into_walkbouts(gps_data, accelerometry_data)
}
\arguments{
\item{gps_data}{GPS data}

\item{accelerometry_data}{Accelerometry data}
}
\value{
A data frame of walk bouts
}
\description{
Convert accelerometry and GPS data into walk bouts
}
