% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gps_data_into_travel_instances.R
\name{process_gps_data_into_travel_instances}
\alias{process_gps_data_into_travel_instances}
\title{Convert GPS data into travel instances}
\usage{
process_gps_data_into_travel_instances(gps_data)
}
\arguments{
\item{gps_data}{GPS data}
}
\value{
A data frame of GPS data, unidentified.
}
\description{
Convert GPS data into travel instances
}
