% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accelerometry_counts_into_bouts.R
\name{run_length_encode}
\alias{run_length_encode}
\title{Run Length Encoding:}
\usage{
run_length_encode(x)
}
\arguments{
\item{x}{a vector to run the function on}
}
\value{
a data.frame with columns for lengths, values, end, and begin
}
\description{
A function that runs a normal run length encoding and adds some extra variables for use in calculations.
}
