% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/identify_walk_bouts_in_gps_and_accelerometry_data.R
\name{summarize_walk_bouts}
\alias{summarize_walk_bouts}
\title{Summarize walking bouts:
This function summarizes walking bouts and calculates the median speed, complete day, non-wearing, bout start, and duration of each bout.}
\usage{
summarize_walk_bouts(walk_bouts, ..., collated_arguments = NULL)
}
\arguments{
\item{walk_bouts}{A data frame containing identified walk bouts}
}
\value{
A data frame summarizing identified walk bouts
}
\description{
Summarize walking bouts:
This function summarizes walking bouts and calculates the median speed, complete day, non-wearing, bout start, and duration of each bout.
}
