% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gps_data_into_gps_epochs.R
\name{validate_gps_data}
\alias{validate_gps_data}
\title{Validate GPS data}
\usage{
validate_gps_data(gps_data)
}
\arguments{
\item{gps_data}{A data frame containing GPS data with the following variables: time, latitude, longitude, and speed.}
}
\value{
This function does not return anything. It throws an error if the GPS data fails any of the validation checks.
}
\description{
This function validates GPS data for required variables, correct variable class, and correct data range.
}
