% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkboutr-package.R
\docType{package}
\name{walkboutr-package}
\alias{walkboutr}
\alias{walkboutr-package}
\title{walkboutr: Generate Walk Bouts from GPS and Accelerometry Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Process GPS and accelerometry data to generate walk bouts. A walk bout is a period of activity with accelerometer movement matching the patterns of walking with corresponding GPS measurements that confirm travel. The inputs of the 'walkboutr' package are individual-level accelerometry and GPS data. The outputs of the model are walk bouts with corresponding times, duration, and summary statistics on the sample population, which collapse all personally identifying information. These bouts can be used to measure walking both as an outcome of a change to the built environment or as a predictor of health outcomes such as a cardioprotective behavior. Kang B, Moudon AV, Hurvitz PM, Saelens BE (2017) \doi{10.1016/j.trd.2017.09.026}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rwalkbout/walkboutr}
  \item \url{https://rwalkbout.github.io/walkboutr/}
  \item Report bugs at \url{https://github.com/rwalkbout/walkboutr/issues}
}

}
\author{
\strong{Maintainer}: Lauren Blair Wilner \email{wilnerl@uw.edu} (\href{https://orcid.org/0000-0003-4439-3734}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Stephen J Mooney \email{sjm2186@uw.edu}
}

}
\keyword{internal}
