% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/identify_walk_bouts_in_gps_and_accelerometry_data.R
\name{identify_walk_bouts_in_gps_and_accelerometry_data}
\alias{identify_walk_bouts_in_gps_and_accelerometry_data}
\title{Convert accelerometry and GPS data into walk bouts}
\usage{
identify_walk_bouts_in_gps_and_accelerometry_data(
  gps_data,
  accelerometry_counts,
  ...,
  collated_arguments = NULL
)
}
\arguments{
\item{gps_data}{GPS data}

\item{accelerometry_data}{Accelerometry data}
}
\value{
A data frame of walk bouts
}
\description{
Convert accelerometry and GPS data into walk bouts
}
