% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{make_active_period}
\alias{make_active_period}
\title{Generate accelerometry datasets}
\usage{
make_active_period(
  length = 1,
  is_bout = TRUE,
  non_wearing = FALSE,
  complete_day = FALSE
)
}
\arguments{
\item{length}{Length of the active period}

\item{is_bout}{Logical indicating if the active period is a bout}

\item{non_wearing}{Logical indicating if the active period is a non-wearing period}

\item{complete_day}{Logical indicating if the active period is a complete day}

\item{active_counts_per_epoch_min}{Minimum active counts per epoch (default: 500)}

\item{minimum_bout_length}{Minimum length of a bout in epochs (default: 10)}

\item{maximum_number_consec_inactive_epochs_in_bout}{Maximum number of consecutive inactive epochs in a bout (default: 3)}

\item{min_non_wearing_length}{Minimum length of non-wearing period in epochs (default: 40)}

\item{min_complete_day}{Minimum length of a complete day in epochs (default: 8602)}

\item{activity_epoch}{A list containing activity counts, bout, non-wearing, and complete day information}
}
\value{
A list of activity epochs
}
\description{
This function generates a list of activity epochs with specified minimum active counts per epoch, minimum bout length,
maximum number of consecutive inactive epochs in a bout, minimum non-wearing length, and minimum complete day length.
}
\examples{
make_active_period()

}
