% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{make_smallest_bout_window}
\alias{make_smallest_bout_window}
\title{Create the smallest bout window}
\usage{
make_smallest_bout_window(minimum_bout_length = 10)
}
\value{
A data.frame with columns \link{activity_counts, bout, non_wearing, complete_day} representing the smallest bout window.
}
\description{
This function creates an active period of minimum length defined by the parameter \code{minimum_bout_length}.
}
