% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{make_smallest_bout_with_largest_inactive_period}
\alias{make_smallest_bout_with_largest_inactive_period}
\title{Generate a sequence of accelerometer counts representing the smallest bout with the largest inactive period}
\usage{
make_smallest_bout_with_largest_inactive_period(
  maximum_number_consec_inactive_epochs_in_bout = 3
)
}
\arguments{
\item{maximum_number_consec_inactive_epochs_in_bout}{maximum number of consecutive inactive epochs in a bout before it is terminated}
}
\value{
A data frame with columns \code{activity_counts} and \code{time}, representing the accelerometer counts and the corresponding time stamps.
}
\description{
This function generates a sequence of accelerometer counts representing the smallest bout with the largest inactive period.
The length of the inactive period is determined by the value of \code{maximum_number_consec_inactive_epochs_in_bout} variable.
}
\examples{
make_smallest_bout_with_largest_inactive_period()
}
