% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{Global parameters and constants}
\format{
An object of class \code{list} of length 5.
}
\usage{
parameters
}
\description{
List of Parameters
\code{epoch_length} The duration of an epoch in seconds.
\code{active_counts_per_epoch_min} Minimum accelerometer counts for an epoch to be considered active (vs. inactive).
\code{minimum_bout_length} Minimum number of epochs for a period of activity to be considered as a potential bout.
\code{local_time_zone} Local time zone of the data - data come in and are returned in UTC, but local time zone is used to compute complete_days.
\code{maximum_number_consec_inactive_epochs_in_bout} Number of consecutive epochs that can be labeled as inactive during a bout without ending the bout.
}
\keyword{datasets}
