% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_accelerometry_counts_into_bouts.R
\name{process_accelerometry_counts_into_bouts}
\alias{process_accelerometry_counts_into_bouts}
\title{Convert accelerometry counts into bouts of physical activity}
\usage{
process_accelerometry_counts_into_bouts(
  accelerometry_counts,
  active_counts_per_epoch_min
)
}
\arguments{
\item{accelerometry_counts}{Raw accelerometry data with the expected schema.}
}
\value{
A data frame of bouts.
}
\description{
The input schema for the accelerometry data is \code{time} and \code{activity_counts}.
\itemize{
\item \code{time} should be a column in date-time format, in the UTC time zone, with no null values.
\item \code{activity_counts} should be a positive numeric column with no null values.
}
}
