% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bouts_and_gps_epochs_into_walkbouts.R
\name{process_bouts_and_gps_epochs_into_walkbouts}
\alias{process_bouts_and_gps_epochs_into_walkbouts}
\title{Convert bouts and gps epochs into walk bouts}
\usage{
process_bouts_and_gps_epochs_into_walkbouts(
  bouts,
  ...,
  collated_arguments = NULL
)
}
\arguments{
\item{bouts}{Accelerometry bouts}

\item{gps_epochs}{GPS epochs}
}
\value{
A data frame of walk bouts and summary statistics.
}
\description{
Convert bouts and gps epochs into walk bouts
}
