% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gps_data_into_gps_epochs.R
\name{process_gps_data_into_gps_epochs}
\alias{process_gps_data_into_gps_epochs}
\title{Convert GPS data into travel instances}
\usage{
process_gps_data_into_gps_epochs(gps_data, ..., collated_arguments = NULL)
}
\arguments{
\item{gps_data}{GPS data}
}
\value{
A data frame of GPS data, unidentified.
}
\description{
The input schema for the accelerometry data is \code{time} and \code{activity_counts}.
\itemize{
\item \code{time} should be a column in date-time format, in the UTC time zone, with no null values.
\item \code{latitude} should be a numeric, non-null latitude coordinate between -90 and 90
\item \code{longitude} should be a numeric, non-null longitude coordinate between -180 and 180
\item \code{speed} should be a numeric, non-null value in kilometers per hour
}
}
