% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_accelerometry_data.R
\name{validate_accelerometry_data}
\alias{validate_accelerometry_data}
\title{Validate accelerometry input data}
\usage{
validate_accelerometry_data(accelerometry_counts)
}
\arguments{
\item{accelerometry_counts}{Raw accelerometry data with the expected schema.}
}
\description{
The input schema for the accelerometry data is \code{time} and \code{activity_counts}.
\itemize{
\item \code{time} should be a column in date-time format, in the UTC time zone, with no null values.
\item \code{activity_counts} should be a positive numeric column with no null values.
}
}
\details{
This function checks the schema of the accelerometry input data
and raises an error if any schema constraints are violated.
}
